update according comments
