+++ superbench/benchmarks/micro_benchmarks/cudnn_function/para_info.json
